Cast size_t to unsigned long to avoid warning.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@22989 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/krb5/pac.c
@@ -158,8 +158,9 @@ krb5_pac_parse(krb5_context context, const void *ptr, size_t len,
 	    goto out;
 	}
 	if (p->pac->buffers[i].offset_lo < header_end) {
-	    krb5_set_error_string(context, "PAC offset inside header: %d %d",
-				  p->pac->buffers[i].offset_lo, header_end);
+	    krb5_set_error_string(context, "PAC offset inside header: %lu %lu",
+				  (unsigned long)p->pac->buffers[i].offset_lo, 
+				  (unsigned long)header_end);
 	    ret = EINVAL;
 	    goto out;
 	}
